+++ internal/tracePersistence/model/scenarioTraceSpan.go
@@ -44,7 +44,7 @@ type Span struct {
 	IsRoot              bool      `json:"is_root"`
 	Kind                string    `json:"kind"`
 	StartTime           time.Time `json:"start_time"`
-	Latency             float64   `json:"latency"`
+	Latency             uint64    `json:"latency"`
 	Source              string    `json:"source"`
 	Destination         string    `json:"destination"`
 	WorkloadIDList      []string  `json:"workload_id_list"`
